Crate yew_macro

This crate provides Yew’s procedural macro html! which allows using JSX-like syntax for generating html and the Properties derive macro for deriving the Properties trait for components.

use yew::prelude::*;

struct Component;

#[derive(Properties, PartialEq)]
struct Props {
    prop: String,
}

// ...

html! {
  <div>
    <button onclick={ctx.link().callback(|_| Msg::Submit)}>
      { "Submit" }
    </button>
    <>
      <Component prop="first" />
      <Component prop="second" />
    </>
  </div>
}

Please refer to https://github.com/yewstack/yew for how to set this up.

Macros

classes

html

html_nested

props

use_prepared_state_with_closure

use_prepared_state_without_closure

use_transitive_state_with_closure

use_transitive_state_without_closure

Attribute Macros

function_component

hook

Derive Macros

Properties